# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AliveDto"}}}}},"tags":["App"]}},"/api/user":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"]}},"/api/user/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}}},"info":{"title":"Artisan API","description":"4 U From Gerste","version":"1.0","contact":{}},"tags":[{"name":"Build with ❤️ by Gerste","description":""}],"servers":[],"components":{"schemas":{"AliveDto":{"type":"object","properties":{"message":{"type":"string"},"status":{"type":"string","enum":["DEVELOPMENT","PRODUCTION"]}},"required":["message","status"]},"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string"},"familyname":{"type":"string"},"email":{"type":"string"},"birthday":{"type":"string","nullable":true},"datevMaNumber":{"type":"number","nullable":true},"sollArbeitszeit":{"type":"number","nullable":true},"wochenarbeitstage":{"type":"number","nullable":true},"urlaubstage":{"type":"number","nullable":true},"eintrittsdatum":{"format":"date-time","type":"string","nullable":true},"austrittsdatum":{"format":"date-time","type":"string","nullable":true},"state":{"type":"string","enum":["active","inactive"],"default":"active"}},"required":["firstname","familyname","email","birthday","datevMaNumber","sollArbeitszeit","wochenarbeitstage","urlaubstage","eintrittsdatum","austrittsdatum","state"]},"User":{"type":"object","properties":{"id":{"type":"string"},"firstname":{"type":"string"},"familyname":{"type":"string"},"email":{"type":"string"},"birthday":{"format":"date-time","type":"string","nullable":true},"datevMaNumber":{"type":"number"},"sollArbeitszeit":{"type":"number"},"wochenarbeitstage":{"type":"number","default":5},"urlaubstage":{"type":"number","default":25},"eintrittsdatum":{"format":"date-time","type":"string"},"austrittsdatum":{"format":"date-time","type":"string"},"state":{"type":"string","enum":["active","inactive"]}},"required":["id","firstname","familyname","email","birthday","datevMaNumber","sollArbeitszeit","wochenarbeitstage","urlaubstage","eintrittsdatum","austrittsdatum","state"]},"UpdateUserDto":{"type":"object","properties":{"firstname":{"type":"string"},"familyname":{"type":"string"},"email":{"type":"string"},"birthday":{"type":"string","nullable":true},"datevMaNumber":{"type":"number"},"sollArbeitszeit":{"type":"number"},"wochenarbeitstage":{"type":"number","default":5},"urlaubstage":{"type":"number","default":25},"eintrittsdatum":{"format":"date-time","type":"string"},"austrittsdatum":{"format":"date-time","type":"string"},"state":{"type":"string","enum":["active","inactive"],"default":"active"}},"required":["firstname","familyname","email"]}}}}